19: Moving ProjectsPreviewItem component to folder
